Rename select-all-controller
